+++ modules/common/guest-platform/src/platform/libraries/stdlib-ranges/kotlin/ranges/IntRange.kt
@@ -8,13 +8,19 @@ package kotlin.ranges
 
 import kotlin.collections.IntIterator
 
-public class IntRange(
+public open class IntProgression(
     public val first: Int,
     public val last: Int,
+    public val step: Int,
 ) {
-    public operator fun contains(value: Int): Boolean = value >= first && value <= last
-
     public external operator fun iterator(): IntIterator
+}
+
+public class IntRange(
+    first: Int,
+    last: Int,
+) : IntProgression(first, last, 1) {
+    public operator fun contains(value: Int): Boolean = value >= first && value <= last
 }
 
 public operator fun Int.rangeTo(other: Int): IntRange = IntRange(this, other)
@@ -22,3 +28,7 @@ public operator fun Int.rangeTo(other: Int): IntRange = IntRange(this, other)
 public external infix fun Int.until(other: Int): IntRange
 
 public external operator fun Int.rangeUntil(other: Int): IntRange
+
+public external infix fun Int.downTo(other: Int): IntProgression
+
+public external infix fun IntProgression.step(step: Int): IntProgression